+++ src/graph/edge/source/edge.cpp
@@ -2,17 +2,17 @@
 
 #include <stdexcept>
 
-edge::edge(size_t weight, vertex* source, vertex* dest) : weight_(weight), source_(source), dest_(dest) {
+edge::edge(long long weight, vertex* source, vertex* dest) : weight_(weight), source_(source), dest_(dest) {
     if (!source || !dest) {
         throw std::invalid_argument("Error: source or dest cannot be nullptr");
     }
 }
 
-void edge::set_weight(size_t new_weight) noexcept {
+void edge::set_weight(long long new_weight) noexcept {
     weight_ = new_weight;
 }
 
-size_t edge::get_weight() const noexcept {
+long long edge::get_weight() const noexcept {
     return weight_;
 }
